improved findNotVerifiedComments

+++ src/main/java/faang/school/postservice/model/AuthorCommentCount.java
@@ -1,6 +1,6 @@
 package faang.school.postservice.model;
 
 public interface AuthorCommentCount {
-    Integer getAuthorId();
+    Long getAuthorId();
     Integer getCount();
 }
